move tour back to navigation; load only on explore variant
