+++ src/components/Navigation/ExploreNavigation/index.js
@@ -1,18 +1,12 @@
 import { Grid, Typography } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
-import dynamic from "next/dynamic";
 import PropTypes from "prop-types";
-import React, { useState } from "react";
+import React from "react";
 
 import SearchIcon from "@/pesayetu/assets/icons/search-explore.svg";
 import DropdownSearch from "@/pesayetu/components/DropdownSearch";
 import Logo from "@/pesayetu/components/Logo";
 import Section from "@/pesayetu/components/Section";
-import { tour } from "@/pesayetu/config";
-
-const Tour = dynamic(() => import("@/pesayetu/components/HURUmap/Tour"), {
-  ssr: false,
-});
 
 const useStyles = makeStyles(({ palette, typography }) => ({
   root: {},
@@ -57,15 +51,10 @@ const useStyles = makeStyles(({ palette, typography }) => ({
   },
 }));
 
-function ExploreNavigation({ logoProps, ...props }) {
+function ExploreNavigation({ logoProps, onOpenHelp, ...props }) {
   const classes = useStyles(props);
-
-  const [isTooltipOpened, setIsTooltipOpened] = useState(false);
-  const handleTooltipOpen = () => {
-    setIsTooltipOpened(true);
-  };
-  const handleTooltipClose = () => {
-    setIsTooltipOpened(false);
+  const openTooltip = () => {
+    onOpenHelp();
   };
   return (
     <div className={classes.root}>
@@ -96,7 +85,7 @@ function ExploreNavigation({ logoProps, ...props }) {
             />
             <Typography
               component="div"
-              onClick={handleTooltipOpen}
+              onClick={openTooltip}
               variant="h3"
               className={classes.help}
             >
@@ -106,9 +95,6 @@ function ExploreNavigation({ logoProps, ...props }) {
           <Grid />
         </Grid>
       </Section>
-      {isTooltipOpened && (
-        <Tour onTooltipClose={handleTooltipClose} {...tour} />
-      )}
     </div>
   );
 }
