+++ src/components/Page/Base.js
@@ -1,16 +1,18 @@
 import { NextSeo } from "next-seo";
+import dynamic from "next/dynamic";
 import PropTypes from "prop-types";
-import React from "react";
+import React, { useState } from "react";
 
 import Footer from "@/pesayetu/components/Footer";
 import Navigation from "@/pesayetu/components/Navigation";
-import { navigationArgs } from "@/pesayetu/config";
+import { navigationArgs, tour } from "@/pesayetu/config";
 import getFooterMenu from "@/pesayetu/functions/menus/getFooterMenu";
 import getNavigationMenu from "@/pesayetu/functions/menus/getNavigationMenu";
 
 /**
  * Base page that can be used to build all other pages.
  */
+let Tour;
 function BasePage({ children, menus, variant, ...props }) {
   const footerProps = getFooterMenu(menus?.footerMenu || []);
   const navigation = getNavigationMenu(menus?.primaryMenu || []);
@@ -23,10 +25,28 @@ function BasePage({ children, menus, variant, ...props }) {
     menuProps,
     socialLinks,
   };
-
+  const [isTooltipOpened, setIsTooltipOpened] = useState(false);
+  const handleTooltipOpen = () => {
+    setIsTooltipOpened(true);
+  };
+  const handleTooltipClose = () => {
+    setIsTooltipOpened(false);
+  };
+  if (variant === "explore") {
+    Tour = dynamic(() => import("@/pesayetu/components/HURUmap/Tour"), {
+      ssr: false,
+    });
+  }
   return (
     <>
-      <Navigation {...navigationProps} variant={variant} />
+      {isTooltipOpened && (
+        <Tour onTooltipClose={handleTooltipClose} {...tour} />
+      )}
+      <Navigation
+        onOpenHelp={handleTooltipOpen}
+        {...navigationProps}
+        variant={variant}
+      />
       <NextSeo {...props} />
       {children}
       {variant !== "explore" && <Footer {...footerProps} />}
